Move Lecturers to officers page
