+++ gatsby/src/pages/officers.js
@@ -18,15 +18,28 @@ export const query = graphql`
         }
       }
     }
+
+    lecturers: allSanityStaff(filter: { lectures: { ne: null } }) {
+      nodes {
+        id
+        name
+        title
+        phone
+        email
+        lectures
+      }
+    }
   }
 `;
 
 const OfficersPage = ({ data }) => {
   const officers = data.officers.nodes[0].officers;
+  const lecturers = data.lecturers.nodes;
 
   return (
     <Layout>
       <OfficersContainer>
+        <h2>Officers</h2>
         {officers.map(({ id, name, title, phone, email }) => {
           return (
             <div className="officers" key={id}>
@@ -34,6 +47,25 @@ const OfficersPage = ({ data }) => {
               <span>{title}</span>
               <span>{phone}</span>
               <span>{email}</span>
+            </div>
+          );
+        })}
+
+        <h3>DDGM/DI</h3>
+        {/* TODO */}
+
+        <h3>Committeemen</h3>
+        {/* TODO */}
+
+        <h3>Lecturers</h3>
+        {lecturers.map(({ id, name, title, phone, email, lectures }) => {
+          return (
+            <div className="lecturer" key={id}>
+              <span>{name}</span>
+              <span>{title}</span>
+              <span>{phone}</span>
+              <span>{email}</span>
+              <span>{lectures}</span>
             </div>
           );
         })}
@@ -47,6 +79,12 @@ const OfficersContainer = styled(Page)`
   flex-wrap: wrap;
 
   .officers {
+    display: flex;
+    flex-direction: column;
+    margin: 20px;
+  }
+
+  .lecturer {
     display: flex;
     flex-direction: column;
     margin: 20px;
